rename poem to synopsis
