+++ build.gradle
@@ -1,64 +1,63 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
 plugins {
 	id 'org.springframework.boot' version '3.1.0'
-	id 'io.spring.dependency-management' version '1.1.0'
-	id 'org.jetbrains.kotlin.jvm' version '1.8.21'
-	id 'org.jetbrains.kotlin.plugin.spring' version '1.8.21'
+    id 'io.spring.dependency-management' version '1.1.0'
+    id 'org.jetbrains.kotlin.jvm' version '1.8.21'
+    id 'org.jetbrains.kotlin.plugin.spring' version '1.8.21'
 }
 
-group = 'open-ai.poem-generator'
+group = 'open-ai.synopsis-generator'
 version = '0.0.1-SNAPSHOT'
 
 java {
-	sourceCompatibility = '17'
+    sourceCompatibility = '17'
+	targetCompatibility = '17'
 }
 
 repositories {
-	mavenCentral()
+    mavenCentral()
 }
 
 springBoot {
-	buildInfo()
+    buildInfo()
 }
 
 dependencyManagement {
-	imports {
-		mavenBom 'org.springframework.cloud:spring-cloud-dependencies:2020.0.4'
-	}
+    imports {
+        mavenBom 'org.springframework.cloud:spring-cloud-dependencies:2022.0.3'
+    }
 }
 
 dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation 'org.springframework.boot:spring-boot-starter-actuator'
-	implementation 'org.springframework.cloud:spring-cloud-starter-bootstrap'
-	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.4'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-actuator'
+    implementation 'org.springframework.cloud:spring-cloud-starter-bootstrap'
+	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.1.0'
 
-	implementation 'com.fasterxml.jackson.module:jackson-module-kotlin'
-	implementation 'org.jetbrains.kotlin:kotlin-reflect'
-	implementation 'com.aallam.openai:openai-client:3.2.5'
-	implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.1")
-	implementation("org.jetbrains.kotlinx:kotlinx-coroutines-reactor:1.7.1")
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-	implementation "com.aallam.openai:openai-client:3.2.5"
-	implementation("io.ktor:ktor-client-java:2.3.1")
-	implementation 'io.reactivex.rxjava2:rxjava:2.1.14'
-	implementation "io.projectreactor:reactor-core:3.5.7"
+    implementation 'com.fasterxml.jackson.module:jackson-module-kotlin'
+    implementation 'org.jetbrains.kotlin:kotlin-reflect'
+    implementation 'com.aallam.openai:openai-client:3.2.5'
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.1")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-reactor:1.7.1")
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    implementation "com.aallam.openai:openai-client:3.2.5"
+    implementation("io.ktor:ktor-client-java:2.3.1")
+    implementation 'io.reactivex.rxjava2:rxjava:2.1.14'
+    implementation "io.projectreactor:reactor-core:3.5.7"
 }
 
 compileKotlin {
-	kotlinOptions {
-		freeCompilerArgs = ['-Xjsr305=strict']
-		jvmTarget = '17'
-	}
+    kotlinOptions {
+        freeCompilerArgs = ['-Xjsr305=strict']
+        jvmTarget = '17'
+    }
 }
 compileTestKotlin {
-	kotlinOptions {
-		freeCompilerArgs = ['-Xjsr305=strict']
-		jvmTarget = '17'
-	}
+    kotlinOptions {
+        freeCompilerArgs = ['-Xjsr305=strict']
+        jvmTarget = '17'
+    }
 }
 
 tasks.named('test') {
-	useJUnitPlatform()
+    useJUnitPlatform()
 }
