+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'poemgenerator'
+rootProject.name = 'synopsis-generator'
