+++ src/main/kotlin/com/openai/synopsisgenerator/controller/SynopsisGeneratorController.kt
@@ -1,6 +1,6 @@
-package com.openai.poemgenerator.controller
+package com.openai.synopsisgenerator.controller
 
-import com.openai.poemgenerator.client.OpenAIClient
+import com.openai.synopsisgenerator.client.OpenAIClient
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/v1/open-ai")
-class PoemGeneratorController(private val serviceClient: OpenAIClient) {
+class SynopsisGeneratorController(private val serviceClient: OpenAIClient) {
 
     @GetMapping("/synopsis")
     suspend fun synopsisGenerator(@RequestParam("actor") actor: String,@RequestParam("genre") genre: String): String {
